build: factor out ModuleInfoException from terasology-module
